fix: store layout::Pages as u32
Saves 4 bytes!!!

+++ crates/msixvc/src/math.rs
@@ -8,21 +8,22 @@ pub fn calculate_hash_block_num_and_run_for_block_num(
     xvd_type: u32,
     hash_tree_levels: u64,
     number_of_hashed_pages: Pages,
-    block_num: u64,
+    block_num: u32,
     hash_level: u32,
     resilient: bool,
     unknown: bool,
-) -> (u64, u64, u64) {
-    fn hash_block_exponent(block_count: u32) -> u64 {
-        (PAGES_PER_BLOCK as u64).pow(block_count)
+) -> (u32, u32, u32) {
+    fn hash_block_exponent(block_count: u32) -> u32 {
+        (PAGES_PER_BLOCK as u32).pow(block_count)
     }
 
     if xvd_type > 1 || hash_level > 3 {
         return (0xFFFF, 0, 1);
     }
 
-    let entry_num_in_block = (block_num / hash_block_exponent(hash_level)) % PAGES_PER_BLOCK as u64;
-    let run_length = PAGES_PER_BLOCK as u64 - entry_num_in_block;
+    let entry_num_in_block =
+        (block_num / hash_block_exponent(hash_level)) % (PAGES_PER_BLOCK as u32);
+    let run_length = PAGES_PER_BLOCK as u32 - entry_num_in_block;
 
     if hash_level == 3 {
         return (0, entry_num_in_block, run_length);
@@ -57,15 +58,15 @@ pub fn calculate_hash_block_num_and_run_for_block_num(
 }
 
 pub fn calculate_number_of_hash_blocks_in_level(
-    size: u64,
+    size: u32,
     hash_level: u64,
     resilient: bool,
-) -> u64 {
+) -> u32 {
     let hash_blocks = match hash_level {
-        0 => size.div_ceil(DATA_BLOCKS_IN_LEVEL0_HASHTREE as u64),
-        1 => size.div_ceil(DATA_BLOCKS_IN_LEVEL1_HASHTREE as u64),
-        2 => size.div_ceil(DATA_BLOCKS_IN_LEVEL2_HASHTREE as u64),
-        3 => size.div_ceil(DATA_BLOCKS_IN_LEVEL3_HASHTREE as u64),
+        0 => size.div_ceil(DATA_BLOCKS_IN_LEVEL0_HASHTREE as u32),
+        1 => size.div_ceil(DATA_BLOCKS_IN_LEVEL1_HASHTREE as u32),
+        2 => size.div_ceil(DATA_BLOCKS_IN_LEVEL2_HASHTREE as u32),
+        3 => size.div_ceil(DATA_BLOCKS_IN_LEVEL3_HASHTREE as u32),
         _ => unreachable!("There are 3 levels"),
     };
 
@@ -76,9 +77,9 @@ pub fn calculate_number_of_hash_blocks_in_level(
     hash_blocks
 }
 
-pub fn calculate_number_of_hash_pages(hashed_pages_count: u64, resilient: bool) -> (u64, u64) {
+pub fn calculate_number_of_hash_pages(hashed_pages_count: u32, resilient: bool) -> (u64, u32) {
     let mut hash_tree_levels = 1;
-    let mut hash_tree_pages = hashed_pages_count.div_ceil(HASH_ENTRIES_IN_PAGE as u64);
+    let mut hash_tree_pages = hashed_pages_count.div_ceil(HASH_ENTRIES_IN_PAGE as u32);
     if hash_tree_pages > 1 {
         let mut result = 2;
         while result > 1 {
